requirement.txt
appnope==0.1.4
asttokens==3.0.0
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
docutils==0.22.2
executing==2.2.1
id==1.5.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
nest-asyncio==1.6.0
nh3==0.3.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
readme_renderer==44.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
twine==6.2.0
urllib3==2.5.0
wcwidth==0.2.13
